requirements.txt
# Production Requirements
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Real-time Features
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# Security & CORS
django-cors-headers==4.3.1
django-filter==23.3
django-ratelimit==4.1.0
cryptography==41.0.7

# File Handling
Pillow==10.1.0
django-imagekit==4.1.0

# Database
psycopg2-binary==2.9.9

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
email-validator==2.1.0
requests==2.31.0
pytz==2023.3
asgiref==3.7.2

# Development Dependencies (install with: pip install -r requirements-dev.txt)